requirements.txt
Flask
google-generativeai
python-dotenv
## requirements to be install in activated environment
